Feat: Mobile 뷰포트의 TypingSection UI를 분리한다
