+++ apps/landing/src/components/main/index.ts
@@ -1,2 +1,4 @@
 export { default as IntroSection } from './IntroSection/IntroSection.component';
 export { default as IntroTypingSection } from './IntroTypingSection/IntroTypingSection.component';
+export { default as IntroTyping } from './IntroTyping/IntroTyping.component';
+export { default as IntroTypingMobile } from './IntroTypingMobile/IntroTypingMobile.component';
